[Add] #98 - FamilyInfo 응답객체 변경, FamilyUser 응답객체 추가
