+++ src/main/java/com/shinhan_hackathon/the_family_guardian/domain/family/dto/FamilyInfoResponse.java
@@ -1,25 +1,11 @@
 package com.shinhan_hackathon.the_family_guardian.domain.family.dto;
 
-import com.shinhan_hackathon.the_family_guardian.domain.user.entity.Level;
-import com.shinhan_hackathon.the_family_guardian.domain.user.entity.Role;
-
 import java.time.LocalDate;
-import java.util.List;
 
 public record FamilyInfoResponse (
         String name,
         String description,
         Integer approvalRequest,
-        List<FamilyUser> userList
+        LocalDate createdAt
 ) {
-
-    public record FamilyUser (
-            Long id,
-            String name,
-            String birthDate,
-            Level level,
-            Role role,
-            String relationship
-    ) {
-    }
 }
