+++ src/main/java/com/shinhan_hackathon/the_family_guardian/domain/family/dto/FamilyUserResponse.java
@@ -1,0 +1,21 @@
+package com.shinhan_hackathon.the_family_guardian.domain.family.dto;
+
+import com.shinhan_hackathon.the_family_guardian.domain.user.entity.Level;
+import com.shinhan_hackathon.the_family_guardian.domain.user.entity.Role;
+
+import java.util.List;
+
+public record FamilyUserResponse(
+        List<FamilyUser> userList
+) {
+
+    public record FamilyUser (
+            Long id,
+            String name,
+            String birthDate,
+            Level level,
+            Role role,
+            String relationship
+    ) {
+    }
+}
